Update README.md
We have used Python software to build our project as it was compatible with
Raspberry pi and its large set of available packages to work with easily.
We have made use of Twilio and IoTSend to broadcast our outputs to the needful
personnel. Apart from these two packages the other big packages used are
numpy(which is used in numerical calculations), SpiDev(which is used for serial
peripheral communication between the Analog to digital converter and Raspberry
pi) and RPi.GPIO (which is used to communicate with the Raspberry pi).